// src/api/model/abridgedFoodItem.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Food Data Central API
 * The FoodData Central API provides REST access to FoodData Central (FDC). It is intended primarily to assist application developers wishing to incorporate nutrient data into their applications or websites.
  To take full advantage of the API, developers should familiarize themselves with the database by reading the database documentation available via links on [Data Type Documentation](https://fdc.nal.usda.gov/data-documentation.html). This documentation provides the detailed definitions and descriptions needed to understand the data elements referenced in the API documentation.
  
  Additional details about the API including rate limits, access, and licensing are available on the [FDC website](https://fdc.nal.usda.gov/api-guide.html)
 * OpenAPI spec version: 1.0.1
 */
import type { AbridgedFoodNutrient } from './abridgedFoodNutrient'

export interface AbridgedFoodItem {
  dataType: string
  description: string
  fdcId: number
  foodNutrients?: AbridgedFoodNutrient[]
  publicationDate?: string
  /** only applies to Branded Foods */
  brandOwner?: string
  /** only applies to Branded Foods */
  gtinUpc?: string
  /** only applies to Foundation and SRLegacy Foods */
  ndbNumber?: number
  /** only applies to Survey Foods */
  foodCode?: string
}